day8 completed

+++ Frontend/lib/logic/services/calculation.dart
@@ -2,8 +2,8 @@
 import 'package:ecommerce/data/models/cart/cart_item_model.dart';
 
 class Calculations{
-  static int cartTotal(List<CartItemModel> items){
-    int total = 0;
+  static double cartTotal(List<CartItemModel> items){
+    double total = 0;
     for (int i = 0; i < items.length; i++){
       total += items[i].product!.price! * items[i].quantity!;
     }
